refactor: consolidate NotFound handling and migrate administrator tabs to dedicated module
