+++ web/src/platform/layouts/Admin.tsx
@@ -3,9 +3,9 @@ import { TopNav } from '@astryxdesign/core/TopNav';
 import { Outlet, useLocation } from 'react-router';
 import { Tab, TabList } from '@astryxdesign/core/TabList';
 import { Auth } from '@/components/Auth';
+import { administratorTabs } from '@/lib/administrator';
 import { findActiveTab } from '@/lib/paths';
 import { ProfileMenu } from '@/components/Profile';
-import { administratorTabs } from '@/platform/tabs';
 import TopLayout from '@/components/layouts/TopLayout';
 import { PageContainer } from '@/components/PageContainer';
 import { useAuthenticatedUser } from '@/lib/hooks/use-user';
